função carrinho

+++ db/migrate/20230117143524_create_produtos.rb
@@ -2,7 +2,7 @@ class CreateProdutos < ActiveRecord::Migration[7.0]
   def change
     create_table :produtos do |t|
       t.string :nome
-      t.integer :preco
+      t.float :preco
 
       t.timestamps
     end
